Clean up frustum code, export from webxr-api
